extract FluentValidation into non-core package
This removes an external reference from the core
functionality of the framework
