+++ CommandDotNet.FluentValidation/CommandDotNet.FluentValidation.csproj
@@ -4,8 +4,8 @@
     <LangVersion>7.3</LangVersion>
     <Authors>Bilal Fazlani, Drew Burlingame</Authors>
     <NeutralLanguage>en-US</NeutralLanguage>
-    <AssemblyTitle>CommandDotNet</AssemblyTitle>
-    <Description>Model your command line application interface in a class</Description>
+    <AssemblyTitle>CommandDotNet.FluentValidation</AssemblyTitle>
+    <Description>Uses FluentValidation validators to validate arguments defined using an IArgumentModel</Description>
     <PackageTags>dotnet core; console; argument parse;</PackageTags>
     <PackageProjectUrl>https://github.com/bilal-fazlani/CommandDotNet</PackageProjectUrl>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
@@ -27,18 +27,6 @@
     <DebugSymbols>true</DebugSymbols>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Humanizer.Core" Version="2.5.16" />
-    <PackageReference Include="LibLog" Version="5.0.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.CSharp" Version="4.5.0" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta2-18618-05" PrivateAssets="All" />
-    <PackageReference Include="System.Reflection" Version="4.3.0" />
-    <PackageReference Include="System.Reflection.Extensions" Version="4.3.0" />
-    <PackageReference Include="System.Reflection.Primitives" Version="4.3.0" />
-  </ItemGroup>
-  <ItemGroup>
     <Compile Remove="output\**" />
   </ItemGroup>
   <ItemGroup>
@@ -48,6 +36,10 @@
     <None Remove="output\**" />
   </ItemGroup>
   <ItemGroup>
-    <None Remove="CommandDotNet.csproj.DotSettings" />
+    <PackageReference Include="FluentValidation" Version="7.6.103" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\CommandDotNet\CommandDotNet.csproj" />
+  </ItemGroup>
+
 </Project>
