+++ CommandDotNet.FluentValidation/ModelValidator.cs
@@ -1,59 +1,25 @@
 ﻿using System;
-using System.Linq;
-using System.Threading.Tasks;
+using System.Reflection;
 using CommandDotNet.Builders;
-using CommandDotNet.Execution;
-using CommandDotNet.Extensions;
 using FluentValidation;
 using FluentValidation.Attributes;
 using FluentValidation.Results;
 
-namespace CommandDotNet.ClassModeling
+namespace CommandDotNet.FluentValidation
 {
-    public class ModelValidator
+    internal class ModelValidator
     {
         // TODO: move FluentValidation into a separate repo & nuget package?
         //       there are other ways to do validation that could also
         //       be applied to parameters
         private readonly IDependencyResolver _dependencyResolver;
 
-        public ModelValidator(IDependencyResolver dependencyResolver)
+        internal ModelValidator(IDependencyResolver dependencyResolver)
         {
             _dependencyResolver = dependencyResolver;
         }
 
-        internal static Task<int> FluentValidationMiddleware(CommandContext commandContext, ExecutionDelegate next)
-        {
-            var modelValidator = new ModelValidator(commandContext.AppConfig.DependencyResolver);
-
-            var paramValues = commandContext.InvocationPipeline
-                .All
-                .SelectMany(i => i.Invocation.ParameterValues.OfType<IArgumentModel>());
-            
-            var failureResults = paramValues
-                .Select(model => new { model, result = modelValidator.ValidateModel(model) })
-                .Where(v => v.result != null && !v.result.IsValid)
-                .ToList();
-
-            if (failureResults.Any())
-            {
-                var console = commandContext.Console;
-                failureResults.ForEach(f =>
-                {
-                    console.Out.WriteLine($"'{f.model.GetType().Name}' is invalid");
-                    foreach (var error in f.result.Errors)
-                    {
-                        console.Out.WriteLine($"  {error.ErrorMessage}");
-                    }
-                });
-                console.Error.WriteLine();
-
-                return Task.FromResult(2);
-            }
-            return next(commandContext);
-        }
-
-        private ValidationResult ValidateModel(IArgumentModel model)
+        internal ValidationResult ValidateModel(IArgumentModel model)
         {
             Type modelType = model.GetType();
