+++ CommandDotNet.Tests/CommandDotNet.Tests.csproj
@@ -17,6 +17,7 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\CommandDotNet.FluentValidation\CommandDotNet.FluentValidation.csproj" />
     <ProjectReference Include="..\CommandDotNet.GitHubReleases\CommandDotNet.NewerReleasesAlerts.csproj" />
     <ProjectReference Include="..\CommandDotNet.TestTools\CommandDotNet.TestTools.csproj" />
     <ProjectReference Include="..\CommandDotNet\CommandDotNet.csproj" />
