+++ CommandDotNet.sln
@@ -23,9 +23,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "scripts", "scripts", "{6296
 		deploy.sh = deploy.sh
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandDotNet.NewerReleasesAlerts", "CommandDotNet.GitHubReleases\CommandDotNet.NewerReleasesAlerts.csproj", "{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CommandDotNet.NewerReleasesAlerts", "CommandDotNet.GitHubReleases\CommandDotNet.NewerReleasesAlerts.csproj", "{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandDotNet.TestTools", "CommandDotNet.TestTools\CommandDotNet.TestTools.csproj", "{75706FAE-BFF4-42FC-A5A6-C49744C7967C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CommandDotNet.TestTools", "CommandDotNet.TestTools\CommandDotNet.TestTools.csproj", "{75706FAE-BFF4-42FC-A5A6-C49744C7967C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandDotNet.FluentValidation", "CommandDotNet.FluentValidation\CommandDotNet.FluentValidation.csproj", "{0E741A7F-AD39-431E-9182-4D337451676C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -45,14 +47,18 @@ Global
 		{73A0E2B7-F205-47E2-A448-7E5CCD7FD6A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{73A0E2B7-F205-47E2-A448-7E5CCD7FD6A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{73A0E2B7-F205-47E2-A448-7E5CCD7FD6A5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB3E8869-107B-4E92-B6A2-2AB3A3CA2D3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{75706FAE-BFF4-42FC-A5A6-C49744C7967C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0E741A7F-AD39-431E-9182-4D337451676C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0E741A7F-AD39-431E-9182-4D337451676C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E741A7F-AD39-431E-9182-4D337451676C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0E741A7F-AD39-431E-9182-4D337451676C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
